+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2 (2024-) - Not for Production
 
 - Add `trim` method
+- Add `replaceBetween` method
 
 ## 0.0.1 (2024-09-02) - Not for Production